Fix doc tests
Includes a workaround for https://github.com/rust-lang/rust/issues/67295

Global context is only available for wasm, but tests are run natively.
Because #[cfg(doctest)] does not work as expected, we include a feature
to fill that role. We should eventually switch back over when the fix is
merged into cargo.