Add support for scoped contexts (#48)
* Refactor context management

* Add root context provider

* Improve docs

* Don't allow global mutation on server side

There are some serious pitfalls using global (thread local) state while
in SSR time. To avoid major security risks, I've added a strict panic if
this is attempted.

Also improve api clarity.

* Remove panic on global write

While this can be a security issue, there are some use cases where it's
perfectly viable to utilize thread-local state on the server. This is
a case where good documentation should help with the niche pitfall,
instead of forbidding it entirely.

* Update docs

* Provide context to store on creation

* Add docs for contexts and srr support

* Adjust dispatch method naming

* Conditionally allow global context access

Only wasm32 targets are allowed global context access. Accessing global
context in a multi-threaded environment is unsafe, and not allowed.

* Fix examples

* Improve docs

* Fix unit tests

* Rename back to Dispatch::global

This does break a lot of code, but is way more descriptive and clear as
to what exactly is happening. Since we are still pre-release, breaking
changes are expected, and shouldn't inhibit improvements.

* Rename with_cx -> new

* Rename Dispatch::cx -> Dispatch::context

* Fix doc tests

Includes a workaround for https://github.com/rust-lang/rust/issues/67295

Global context is only available for wasm, but tests are run natively.
Because #[cfg(doctest)] does not work as expected, we include a feature
to fill that role. We should eventually switch back over when the fix is
merged into cargo.

* Update docs

* Fix macro

Also remove unecessary derives

* Improve context docs

* Refactor functional hooks slightly

* Improve docs