Fix listener recursion
Fixes a panic when trying to reduce a store from the listener of that
same store. To accomplish this we had to restrict Listener::on_change
to be immutable borrow of self only. Although this means we can no
longer keep local state in listener's, allowing recursion and avoiding
the panic seems an improvement.

Now to track and manage separate state, we need only create a new store.